pyproject.toml
[project]
name = "hotspot-socks-proxy"
version = "3.0.3"
description = "SOCKS proxy for routing traffic through WiFi interface"
authors = [
    { name = "tadeasf", email = "business@tadeasfort.com" }
]
dependencies = [
    "typer>=0.15.1",
    "prompt-toolkit>=3.0.48",
    "rich>=13.9.4",
    "psutil>=5.9.0",
    "pyperclip>=1.9.0",
    "tomli>=2.2.1",
    "dnspython>=2.7.0",
    "mkdocs-gen-files>=0.5.0",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.build]
includes = ["src/hotspot_socks_proxy"]
package-dir = "src"

[project.scripts]
hotspot-proxy = "hotspot_socks_proxy.cmd.cli:app"

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D203",  # no-blank-line-before-class
    "UP036", # unused-import
    "RUF022", # unsorted-dunder-all
    "D212",  # multi-line-summary-first-line
    "ANN001", # missing-type-self
    "ANN002", # missing-type-cls
    "ANN201", # missing-type-return
    "ANN401", # missing-type-annotation
    "BLE001", # blacklisted-import
    "TRY300", # try-except-pass
    "COM812", # missing trailing comma
    "ISC001", # single line import statements
    "S603",   # subprocess call security
    "C901",   # function too complex
]

[tool.ruff.lint.isort]
known-first-party = ["hotspot_socks_proxy"]
combine-as-imports = true

[tool.ruff.lint.pydocstyle]
convention = "google"
